extractionsite tooltip basics
